Remove normalize_parameters function.
This function was actually decoding parameters that shouldn't have been
decoded, then re-encoding them.

See: https://github.com/WP-API/OAuth1/issues/91

The idea behind removing this function is that parameters should be
encoded as needed, not through a brute-force function like this, which
can cause unintended side-effects for parameters which contain portions
that are already encoded.