Properly show verification UI during out-of-band flow
The conditional in this function makes it so that a callback of "oob"
will never trigger the UI view that allows a user to manually copy
their verification code for out-of-band OAuth flow.

This commit does not necessarily resolve #149, which is about the OOB
UI showing up when it's not supposed to; but at least this does make
OOB verification work in my test environment.